First Commit.
This is as far as I could get with this. Was hoping that the switch
would append colored boxes as it ran through. I looked at and tried a
number of different way to get to this. I had a random color generator,
but couldn’t figure out how the box would know what color it was. I
can’t work on this anymore today (an actual date, no kids with the
wife, Sigur Ros at the Orpheum!!). I’d like someone to talk a few
things out with the code if possible.